+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 5.0.0-rc.next
 - Adds the `Notifier` API to `VcsUiApp`.
 - Adds a `VcsList` component for standard item list rendering.
+- Adds new `VcsWizard` component as stylized wrapper of Vuetify VStepper
 
 # 5.0.0-rc.15
 - Update Openlayers to version 7.0.0